Refactor EC pannel settings.
