+++ js/src/components/enhanced-conversion-tracking-settings/toggle.js
@@ -2,7 +2,12 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { useCallback, useEffect } from '@wordpress/element';
+import {
+	useCallback,
+	useEffect,
+	Fragment,
+	createInterpolateElement,
+} from '@wordpress/element';
 import { Spinner } from '@woocommerce/components';
 
 /**
@@ -11,8 +16,9 @@ import { Spinner } from '@woocommerce/components';
 import { ENHANCED_ADS_CONVERSION_STATUS } from '.~/constants';
 import { useAppDispatch } from '.~/data';
 import AppStandaloneToggleControl from '.~/components/app-standalone-toggle-control';
+import TrackableLink from '.~/components/trackable-link';
 import useAllowEnhancedConversions from '.~/hooks/useAllowEnhancedConversions';
-import useAcceptedCustomerDataTerms from '.~/hooks/useAcceptedCustomerDataTerms';
+import useGoogleAdsEnhancedConversionSettingsURL from '.~/hooks/useGoogleAdsEnhancedConversionSettingsURL';
 
 const TOGGLE_LABEL_MAP = {
 	[ ENHANCED_ADS_CONVERSION_STATUS.DISABLED ]: __(
@@ -26,22 +32,10 @@ const TOGGLE_LABEL_MAP = {
 };
 
 const Toggle = () => {
-	const { updateEnhancedAdsConversionStatus, invalidateResolution } =
-		useAppDispatch();
+	const { url } = useGoogleAdsEnhancedConversionSettingsURL();
+	const { updateEnhancedAdsConversionStatus } = useAppDispatch();
 	const { allowEnhancedConversions, hasFinishedResolution } =
 		useAllowEnhancedConversions();
-	const {
-		acceptedCustomerDataTerms,
-		hasFinishedResolution: hasResolvedAcceptedCustomerDataTerms,
-	} = useAcceptedCustomerDataTerms();
-
-	useEffect( () => {
-		if (
-			allowEnhancedConversions === ENHANCED_ADS_CONVERSION_STATUS.PENDING
-		) {
-			invalidateResolution( 'getAcceptedCustomerDataTerms', [] );
-		}
-	}, [ allowEnhancedConversions, invalidateResolution ] );
 
 	const handleOnChange = useCallback(
 		( value ) => {
@@ -59,18 +53,38 @@ const Toggle = () => {
 	}
 
 	return (
-		<AppStandaloneToggleControl
-			checked={
-				allowEnhancedConversions ===
-				ENHANCED_ADS_CONVERSION_STATUS.ENABLED
-			}
-			disabled={ ! hasResolvedAcceptedCustomerDataTerms }
-			onChange={ handleOnChange }
-			label={
-				TOGGLE_LABEL_MAP[ allowEnhancedConversions ] ||
-				__( 'Enable', 'google-listings-and-ads' )
-			}
-		/>
+		<Fragment>
+			<p>
+				{ createInterpolateElement(
+					__(
+						'To properly measure your enhanced conversions, please ensure Enhanced Conversions is enabled in <link>Google Ads Settings</link>, and that the setup method chosen is Google Tag',
+						'google-listings-and-ads'
+					),
+					{
+						link: (
+							<TrackableLink
+								target="_blank"
+								type="external"
+								href={ url }
+							/>
+						),
+					}
+				) }
+			</p>
+
+			<AppStandaloneToggleControl
+				checked={
+					allowEnhancedConversions ===
+					ENHANCED_ADS_CONVERSION_STATUS.ENABLED
+				}
+				disabled={ ! hasFinishedResolution }
+				onChange={ handleOnChange }
+				label={
+					TOGGLE_LABEL_MAP[ allowEnhancedConversions ] ||
+					__( 'Enable', 'google-listings-and-ads' )
+				}
+			/>
+		</Fragment>
 	);
 };
